+++ mongox/codec_test.go
@@ -39,6 +39,21 @@ var _ = Describe("Codec", func() {
 				err := ProtoDeCodecFunc(bsoncodec.DecodeContext{}, &MockbsonReader{ErrReadDateTime: errors.New("error read time")}, reflect.ValueOf(ts))
 				Expect(err).NotTo(Succeed())
 			})
+			It("str codec", func() {
+				ts := "2022-01-01T00:00:00Z"
+				err := StrCodecFunc(bsoncodec.EncodeContext{}, &MockbsonWriter{}, reflect.ValueOf(ts))
+				Expect(err).To(Succeed())
+			})
+			It("str codec wrong input type", func() {
+				ts := 1
+				err := StrCodecFunc(bsoncodec.EncodeContext{}, &MockbsonWriter{}, reflect.ValueOf(ts))
+				Expect(err).NotTo(Succeed())
+			})
+			It("str codec wrong ts format", func() {
+				ts := "hello world"
+				err := StrCodecFunc(bsoncodec.EncodeContext{}, &MockbsonWriter{}, reflect.ValueOf(ts))
+				Expect(err).To(Succeed())
+			})
 		})
 	})
 })